Refactor API endpoints: modularize analytics, content, and scheduling routes; enhance post generation and scheduling logic
